Fixed comments: moved contexts to separate files, removed unnecessary useCallback invocations
