+++ src/components/MovieListCard/MovieListCard.tsx
@@ -7,10 +7,10 @@ import { Movie } from '../../models/Movie';
 import Modal from '../Modal/Modal';
 import DeleteMovieConfirm from '../DeleteMovieConfirm/DeleteMovieConfirm';
 import EditMovieForm from '../EditMovieForm/EditMovieForm';
-import { SelectedMovieContext } from '../../App';
 import { getYear } from '../../utils/getYearFromDate';
 import { joinGenres } from '../../utils/joinGenresWithComma';
 import React from 'react';
+import { SelectedMovieContext } from '../../contexts/SelectedMovieContext';
 
 interface MoviesListCardProps {
   movie: Movie;
@@ -61,8 +61,8 @@ function MoviesListCard({ movie }: MoviesListCardProps) {
     console.log(formValue);
   }, []);
 
-  const closeEditMovieModal = useCallback(() => setMovieToEdit(null), []);
-  const closeDeleteMovieModal = useCallback(() => setMovieToDelete(null), []);
+  const closeEditMovieModal = () => setMovieToEdit(null);
+  const closeDeleteMovieModal = () => setMovieToDelete(null);
 
   const memoizedYear = useMemo(() => getYear(release_date), [release_date]);
   const memoizedGenres = useMemo(() => joinGenres(genres), [genres]);
