+++ src/contexts/FetchedMoviesContext.ts
@@ -1,0 +1,7 @@
+import React from 'react';
+import { UseFetchMoviesResult } from '../hooks/useFetchMovies';
+
+export const FetchedMoviesContext = React.createContext<UseFetchMoviesResult>([
+  { fetchedMovies: [], isError: false, isLoading: false, queryParams: {} },
+  () => {},
+]);
